Haier_AI_Product_Info_Chatbot/requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.3
llama-index==0.12.10
llama-index-core==0.12.10.post1
llama-index-embeddings-gemini==0.3.1
llama-index-llms-gemini==0.4.3